split EntityFactory into different Factories
